fix(init): Move configuration finalization out of Main
